+++ pages/index.tsx
@@ -10,7 +10,7 @@ export default function Home() {
             </Head>
             <LinkWrapper>
                 <Link href="/folder">FolderPage 이동하기</Link>
-                <Link href="/shared">Sharedpage 이동하기</Link>
+                <Link href="/shared/41">Sharedpage 이동하기</Link>
                 <Link href="/signin">SigninPage 이동하기</Link>
                 <Link href="/signup">SignupPage 이동하기</Link>
             </LinkWrapper>
