+++ pages/shared/[folderId]/index.tsx
@@ -1,12 +1,12 @@
 import { useSearchBar } from "@/hooks/useSearchBar";
 import { useGetShareLinkList } from "@/hooks/Shared.hook";
-import Header from "@/components/Header/Header";
-import ShareDescription from "@/components/ShareDescription/ShareDescription";
-import Contents from "@/components/Contents/Contents";
-import CardSearchBar from "@/components/Contents/CardSearchBar/CardSearchBar";
-import Footer from "@/components/Footer/Footer";
-import CardList from "@/components/Contents/CardList/CardList";
 import { useRouter } from "next/router";
+import Header from "@/components/Common/Header/Header";
+import Contents from "@/components/Common/Contents/Contents";
+import CardSearchBar from "@/components/Common/Contents/CardSearchBar/CardSearchBar";
+import CardList from "@/components/Common/Contents/CardList/CardList";
+import Footer from "@/components/Common/Footer/Footer";
+import Description from "@/components/SharedPage/Description";
 
 export default function SharedFolderDetail() {
     const router = useRouter();
@@ -26,7 +26,7 @@ export default function SharedFolderDetail() {
     return (
         <>
             <Header currentPath="shared" />
-            <ShareDescription />
+            <Description />
             <Contents>
                 <CardSearchBar
                     inputValue={inputValue}
